pyproject.toml
[tool.poetry]
name = "backend-tg"
version = "0.1.0"
description = ""
authors = ["ViniciusEZ <ez998@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
python-decouple = "^3.8"
django-cors-headers = "^4.5.0"
dj-database-url = "^2.3.0"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
tqdm = "^4.67.0"
django-db-connection-pool = "^1.2.5"
adrf = "^0.1.8"
django-redis = "^5.4.0"
locust = "^2.32.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
